Computational Geometry(Python)/SegmentIntersection/runtime_demo/utils/smart_file/callback_that_uses_file.py
from typing import Any, TextIO
from interface import Interface  # type: ignore


class CallbackThatUsesFile(Interface):
    def run(self, file_handler: TextIO = None, **kwargs) -> Any:
        pass



